refactor: Update Firestore rules to restrict task access based on user authentication

+++ src/services/tasks.ts
@@ -7,7 +7,7 @@ import {
   getDocs,
   updateDoc,
 } from "firebase/firestore"
-import { db } from "../firebase"
+import { auth, db } from "../firebase"
 import type { Moment } from "moment"
 import type { Label } from "./label"
 
@@ -24,6 +24,12 @@ export interface CreateTask {
 export const taskCollection = collection(db, "tasks")
 
 export async function createTask(task: CreateTask) {
+  const user = auth.currentUser
+
+  if (!user) {
+    throw new Error("User not found")
+  }
+
   const dueDate = task.dueDate.toDate()
   const isCompleted = false
   const labels = task.labels?.filter(label => label.id).map(label => label.id)
@@ -38,16 +44,34 @@ export async function updateTask(taskId: string, task: Partial<CreateTask>) {
 }
 
 export async function getTasks() {
+  const user = auth.currentUser
+
+  if (!user) {
+    throw new Error("User not found")
+  }
+
   const taskSnapshot = await getDocs(taskCollection)
   const tasks = taskSnapshot.docs.map(doc => doc.data())
   return tasks
 }
 
 export async function deleteTask(taskId: string) {
+  const user = auth.currentUser
+
+  if (!user) {
+    throw new Error("User not found")
+  }
+
   await deleteDoc(doc(db, "tasks", taskId))
 }
 
 export async function toggleTaskCompletion(taskId: string) {
+  const user = auth.currentUser
+
+  if (!user) {
+    throw new Error("User not found")
+  }
+
   const taskDoc = doc(db, "tasks", taskId)
   const task = await getDoc(taskDoc)
   await updateDoc(taskDoc, { isCompleted: !task.data()!.isCompleted })
